donation firebase functionality

+++ app/src/main/java/com/example/rahul/donationtrackerapp/Model/Item.java
@@ -25,7 +25,7 @@ public class Item {
         this.location = location;
     }
 
-    public double getKey() {
+    public int getKey() {
         return key;
     }
